refresh feature upload and bug fix 1
