+++ teamwork/apps/profiles/admin.py
@@ -23,6 +23,6 @@ class UserAdmin(BaseUserAdmin):
 admin.site.unregister(User)
 admin.site.register(User, UserAdmin)
 admin.site.register(Skills)
-admin.site.register(Techs)
+# admin.site.register(Techs)
 admin.site.register(Profile)
 admin.site.register(Alert)
