+++ teamwork/apps/profiles/views/EditProfileView.py
@@ -31,23 +31,6 @@ def edit_skills(request, username):
 
     return HttpResponse("Failure")
 
-def edit_techs(request, username):
-    if request.method == 'GET' and request.is_ajax():
-        # JSON prefers dictionaries over lists.
-        data = dict()
-        # A list in a dictionary, accessed in select2 ajax
-        data['items'] = []
-        q = request.GET.get('q')
-        if q is not None:
-            results = Techs.objects.filter(
-                Q( tech__contains = q ) ).order_by( 'tech' )
-        for s in results:
-            data['items'].append({'id': s.tech, 'text': s.tech})
-        return JsonResponse(data)
-
-
-    return HttpResponse("Failure")
-
 @login_required
 def edit_profile(request, username):
     """
@@ -181,9 +164,8 @@ def edit_profile_helper(request, username):
     else:
         #grab profile for the current user
         profile = Profile.objects.get(user=request.user)
-        tempProfile = User.objects.get(username=username)
+        profileUser = User.objects.get(username=profile.user)
         #grab profile for the current user
-
 
     #request.FILES is passed for File storing
     form = ProfileForm(request.POST, request.FILES)
@@ -201,7 +183,7 @@ def edit_profile_helper(request, username):
         if name:
             profile.name = name
         if email:
-            tempProfile.email=email
+            profileUser.email = email
         if bio:
             profile.bio = bio
         if institution:
@@ -212,4 +194,4 @@ def edit_profile_helper(request, username):
             profile.avatar = ava
 
         profile.save()
-        tempProfile.save()
+        profileUser.save()
