+++ teamwork/apps/projects/admin.py
@@ -7,10 +7,8 @@ from .models import Interest, Membership, Project, ProjectUpdate, ResourceUpdate
 # Register the following models for the admin site
 admin.site.register(Project)
 
-admin.site.unregister(Techs)
+# admin.site.unregister(Techs)
 admin.site.register(Techs)
-
-
 
 admin.site.register(Membership)
 admin.site.register(ProjectUpdate)
