+++ teamwork/templates/adminlte/lib/_main_header.html
@@ -115,25 +115,25 @@
                             <div class="btn-group">
                               <a href="{% url 'profile' user.username %}">
 							  <!-- need to add margin:2px 0 for Light/Dark theme -->
-                                  <button type="button" class="btn btn-flat" style="background: rgba(243, 156, 18, 1);margin:2px 0">
+                                  <button type="button" class="btn btn-default" style="background: rgba(243, 156, 18, 1);margin:2px 0">
                                       <font color="white">Profile</font>
                                   </button>
                               </a>
 							  <!-- need to add margin:2px 0 for Light/Dark theme -->
                               <a href="{% url 'edit_schedule' user.username %}">
-                                  <button type="button" class="btn btn-flat" style="background: rgba(0, 150, 255, 1);margin:2px 0">
+                                  <button type="button" class="btn btn-default" style="background: rgba(0, 150, 255, 1);margin:2px 0">
                                       <font color="white">Edit Schedule</font>
                                   </button>
                               </a>
 							  <!-- main code of the Light/Dark theme switch -->
 							  <a href="javascript:switchSkin()">
-                                  <button type="button" class="btn btn-flat  btn-info" style="margin:2px 0">
+                                  <button type="button" class="btn btn-default btn-info" style="margin:2px 0">
                                       <font color="white">Switch Theme</font>
                                   </button>
                               </a>
 							  <!-- add style="display:block;clear:both;" and margin:2px 0;width:100%; for Light/Dark theme -->
                               <a href="{% block logout_url %}{% logout_url %}{% endblock %}"style="display:block;clear:both;">
-                                  <button type="button" class="btn btn-flat" style="background: rgba(255, 0, 0, 0.7);margin:2px 0;width:100%;">
+                                  <button type="button" class="btn btn-default" style="background: rgba(255, 0, 0, 0.7);margin:2px 0;width:100%;">
                                       <font color="white">Sign Out</font>
                                   </button>
                               </a>
