+++ teamwork/templates/projects/partial_project_profile.html
@@ -62,7 +62,7 @@
     {% if forloop.last %}</div><div class="clearfix"></div>{% endif %}
 {% empty %}
   <div>
-  </h3>They are not in any projects</h3>
+  </h3>You are not in any projects</h3>
 
   </div>
 {% endfor %}
